requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.5.3
black==19.10b0
chardet==3.0.4
click==7.1.2
colorthief==0.2.1
config==0.5.0.post0
cr-py==1.2.7
discord.py==1.4.0a2293+g874c2cd
python-dotenv==0.12.0